decouple schema processing from model
Moves the processing logic for iterating over the OpenAPI3 schema
elements into a pkg/schema package, creating a Helper struct that wraps
nicely all the funky logic for constructing the model information for
the API itself.

"Resource" has been renamed to "CRD" to better represent what it is.
