+++ pkg/schema/enum_def.go
@@ -11,35 +11,21 @@
 // express or implied. See the License for the specific language governing
 // permissions and limitations under the License.
 
-package resource
+package schema
 
 import (
-	"bytes"
 	"fmt"
 
-	"github.com/getkin/kin-openapi/openapi3"
-
+	"github.com/aws/aws-service-operator-k8s/pkg/model"
 	"github.com/aws/aws-service-operator-k8s/pkg/names"
 )
 
-type EnumValue struct {
-	Original string
-	Clean    string
-}
-
-type EnumDef struct {
-	Names  names.Names
-	GoType string
-	Values []EnumValue
-}
-
-func EnumDefsFromAPI(
-	api *openapi3.Swagger,
-) ([]*EnumDef, error) {
-	edefs := []*EnumDef{}
+func (h *Helper) GetEnumDefs() ([]*model.EnumDef, error) {
+	api := h.api
+	edefs := []*model.EnumDef{}
 
 	for schemaName, schemaRef := range api.Components.Schemas {
-		schema := getSchemaFromSchemaRef(api, schemaRef)
+		schema := h.getSchemaFromSchemaRef(schemaRef)
 		if len(schema.Enum) == 0 {
 			continue
 		}
@@ -57,35 +43,19 @@ func EnumDefsFromAPI(
 		default:
 			return nil, fmt.Errorf("cannot determine go type from enum schema type %s", schema.Type)
 		}
-		vals := make([]EnumValue, len(schema.Enum))
+		vals := make([]model.EnumValue, len(schema.Enum))
 		for x, item := range schema.Enum {
 			strVal, ok := item.(string)
 			if !ok {
 				return nil, fmt.Errorf("cannot convert %v to string", item)
 			}
-			vals[x] = newEnumVal(strVal)
+			vals[x] = model.NewEnumVal(strVal)
 		}
-		edefs = append(edefs, &EnumDef{
+		edefs = append(edefs, &model.EnumDef{
 			Names:  names.New(schemaName),
 			GoType: goType,
 			Values: vals,
 		})
 	}
 	return edefs, nil
-}
-
-func newEnumVal(orig string) EnumValue {
-	// Convert values like "m5.xlarge" into "m5_xlarge"
-	cleaner := func(r rune) rune {
-		if (r >= 'A' && r <= 'Z') || (r >= 'a' && r <= 'z') || (r >= '0' && r <= '9') {
-			return r
-		}
-		return '_'
-	}
-	clean := bytes.Map(cleaner, []byte(orig))
-
-	return EnumValue{
-		Original: orig,
-		Clean:    string(clean),
-	}
 }
